A new module was created to save the constants
