+++ CaptureTheFlag/CaptureTheFlag/Command/CmdPublic.cs
@@ -10,6 +10,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using static CaptureTheFlag.GameMode;
+using CaptureTheFlag.Constants;
 
 namespace CaptureTheFlag.Command
 {
