+++ CaptureTheFlag/CaptureTheFlag/Constants/FlagID.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CaptureTheFlag.Constants
+{
+    public class FlagID
+    {
+        public static int Alpha { get; } = 19306;
+        public static int Beta { get; } = 19307;
+    }
+}
