+++ CaptureTheFlag/CaptureTheFlag/Constants/SkinTeam.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CaptureTheFlag.Constants
+{
+    public class SkinTeam
+    {
+        public static int Alpha { get; } = 170;
+        public static int Beta { get; } = 177;
+    }
+}
