+++ CaptureTheFlag/CaptureTheFlag/Constants/StateUser.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CaptureTheFlag.Constants
+{
+    public enum StateUser
+    {
+        Force,
+        Kill,
+        None
+    }
+}
