+++ CaptureTheFlag/CaptureTheFlag/Constants/TeamID.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace CaptureTheFlag.Constants
+{
+    public enum TeamID
+    {
+        Alpha,
+        Beta
+    }
+}
