+++ CaptureTheFlag/CaptureTheFlag/Flag.cs
@@ -64,10 +64,4 @@ namespace CaptureTheFlag
                 pickup.Dispose();
         }
     }
-
-    public class FlagID
-    {
-        public static int Alpha { get; } = 19306;
-        public static int Beta { get; } = 19307;
-    }
 }
