+++ CaptureTheFlag/CaptureTheFlag/Map/CurrentMap.cs
@@ -10,6 +10,7 @@ using static CaptureTheFlag.Rand;
 using static CaptureTheFlag.GameMode;
 using static SampSharp.GameMode.World.BasePlayer;
 using static CaptureTheFlag.Map.FileRead;
+using CaptureTheFlag.Constants;
 
 namespace CaptureTheFlag.Map
 {
